#112 Refactored code layout
