+++ .github/workflows/go.yml
@@ -7,7 +7,7 @@ on:
     branches: [ main ]
 
 jobs:
-  build:
+  build-atc:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
@@ -18,10 +18,10 @@ jobs:
         go-version: 1.13
 
     - name: Build
-      run: make build
+      run: cd ATC; make build
 
     - name: Test
-      run: make test
+      run: cd ATC; make test
       env:
         CAPTAIN_DRY_RUN: TRUE
         CAPTAIN_DB: testing.db
@@ -31,12 +31,12 @@ jobs:
       env:
         CC_TEST_REPORTER_ID: c19b466b4c3e08a48f23e2c883a52c68bc7b2a3b2f52ab914224004b5a990df3
       with:
-        prefix: github.com/ARMmaster17/Captain
-        coverageLocations: ${{github.workspace}}/cover.out:gocov
+        prefix: github.com/ARMmaster17/Captain/ATC
+        coverageLocations: ${{github.workspace}}/ATC/cover.out:gocov
 
     - name: Publish unit test results
       uses: EnricoMi/publish-unit-test-result-action@v1
       if: always()
       with:
-        files: report.xml
+        files: ./ATC/report.xml
         report_individual_runs: "true"
