+++ ATC/DB/DB_test.go
@@ -1,4 +1,4 @@
-package db
+package DB
 
 import (
 	"os"
