+++ ATC/StateServer_test.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	db2 "github.com/ARMmaster17/Captain/db"
+	db2 "github.com/ARMmaster17/Captain/ATC/DB"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
